% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ChIPdata-class}
\alias{ChIPdata-class}
\alias{ChIPdata}
\alias{ChIPdata,}
\alias{ChIPdata-class}
\title{ChIPdata object and constructors}
\usage{
ChIPdata(bamfile = NULL, reads = NULL, isPE = NULL)
}
\arguments{
\item{bamfile}{a string indicating the location of the aligned reads bamfile}

\item{reads}{either a \code{GAlignments} or a \code{GAlignmentPairs} object with
the aligned reads of the ChIP experiment}

\item{isPE}{a logical value indicating if the reads are Single-End (SE) or Paired-End (PE). 
In the case that both \code{reads} and \code{isPE} are both used, then the \code{isPE} value
will be determined from \code{reads} and raise a warning}
}
\value{
a \code{ChIPdata} object
}
\description{
\code{ChIPdata} is used to contain either SE or PE reads, and the idea
is to contain some of the most commonly used functions in a 
ChIP-seq/exo/nexus analysis.
}
\examples{
file <- system.file("extdata","example","encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
ChIPdata(bamfile = file,isPE = FALSE)

reads <- readGAlignments(file,param = NULL)
ChIPdata(reads = reads)
}
