% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_bins.R
\name{create_bins}
\alias{create_bins}
\title{Partitions a genome into bins of a fixed length and counts the number of reads that overlap each bin}
\usage{
create_bins(bin_size, reads = NULL, chrom = NULL, frag_len = 1)
}
\arguments{
\item{bin_size}{A integer value used to partition the chromosome}

\item{reads}{A reads object}

\item{chrom}{A GRanges object specifying the genome to bin. The maximum length used to create the bins
is gonna be used the integer part of (chromLen / fragLen) times fragLen for each chromosome.}

\item{frag_len}{An integer value used to extend the fragments. The default value is one, to count only the 5' ends
that overlaps the bins}
}
\value{
A GRanges object. If a reads object is present then is going to count the number of extended fragments per
bin
}
\description{
Partitions a chromosome into fixed length bins and counts the number of extended fragments that
overlap each bin
}
\examples{
file <- system.file("extdata","example",
  "encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
reads <- create_reads(file)
chrom <- GRanges(seqnames = "chr1", ranges = IRanges(start = 1,end = 249250621))
create_bins(2000,reads = reads)
create_bins(2000, chrom = chrom )
create_bins(2000, reads = reads, frag_len = 200)
create_bins(2000, reads = reads,chrom = chrom, frag_len = 200)
}

