% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_reads.R
\name{create_reads}
\alias{create_reads}
\title{Creates a reads object}
\usage{
create_reads(readsFile)
}
\arguments{
\item{readsFile}{Name of the bam file with the aligned reads}
}
\value{
A reads object
}
\description{
Constructor of the reads class
}
\examples{
\dontrun{
file = ""
create_reads(file)
}
}

