% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt2gr}
\alias{dt2gr}
\title{Converts a data.table to GRanges}
\usage{
dt2gr(dt)
}
\arguments{
\item{dt}{A data.table with at least the columns seqnames,start and end}
}
\value{
A GRanges object
}
\description{
Coverts a data.table into a GRanges object
}
\examples{

dt <- data.table(seqnames = rep("chr1",8),start = 3:10,end = sample(11:18,8))
dt2gr(dt)

dt2 <- copy(dt); dt2[,extra := runif(8)]
dt2gr(dt2)

dt3 <- copy(dt); dt3[,strand := sample(c("+","-"),8,replace = TRUE)]
dt2gr(dt3)

dt3[ , extra := runif(8)]
dt2gr(dt3)
}
\seealso{
\code{\link{gr2dt}}
}

