% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods_chipdata.R
\docType{methods}
\name{isPET}
\alias{isPET}
\alias{isPET,ChIPdata-method}
\alias{isPET}
\title{isPET methods}
\usage{
isPET(object)

\S4method{isPET}{ChIPdata}(object)
}
\arguments{
\item{object}{A \code{ChIPdata} object}
}
\value{
A logical value indicating if the reads are PET (TRUE) or SET (FALSE)
}
\description{
isPET returns a logical value indicating if the object isPET
}
\seealso{
\code{\link{ChIPdata-class}}
}
