% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_strand_cross_corr.R
\name{local_strand_cross_corr}
\alias{local_strand_cross_corr}
\title{Calculate the strand cross correlation for a specific genome region}
\usage{
local_strand_cross_corr(reads, region, shift = 1:300)
}
\arguments{
\item{reads}{A reads object}

\item{region}{A GRanges object with length 1}

\item{shift}{An integer vector, the default value is \code{1:300}}
}
\value{
A data.table with two columns \code{shift} and \code{cross.corr}
}
\description{
Calculates the strand cross correlation as defined by
Kharchenko et. al. 2008 conditional to a specific region in the genome
}
\examples{
file <- system.file("extdata","example",
  "encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
reads <- create_reads(file)
region <- GRanges(seqnames = "chr1", ranges = IRanges(start = 1,end = 249250621))
local_strand_cross_corr(reads,region)
}

