% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_bins.R
\name{MAPlot}
\alias{MAPlot}
\title{Creates a quick MA plot to check the relationship between two reads objects}
\usage{
MAPlot(x, y, binSize, nrBins = 40, chrom = NULL, fragLen = 1)
}
\arguments{
\item{x}{A \code{ChIPdata} object that corresponds to the x-axis}

\item{y}{Another \code{ChIPdata} object, this one correspond to the y-axis}

\item{binSize}{A integer value used to partition the chromosome}

\item{nrBins}{Integer value with the number of bins used to build the plot. The default value is 100}

\item{chrom}{A GRanges object specifying the genome to bin. The maximum length used to create the bins
is gonna be used the integer part of (chromLen / fragLen) times fragLen for each chromosome.}

\item{fragLen}{An integer value used to extend the fragments. The default value is one, to count only the 5' ends 
that overlaps the bins}
}
\value{
A ggplot object with the hexbin plot
}
\description{
Creates a quick MA plot to visualize the relationship between 
two reads objects
}
