% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods_chipdata.R
\docType{methods}
\name{nreads}
\alias{nreads}
\alias{nreads,ChIPdata-method}
\alias{nreads}
\title{nreads methods}
\usage{
nreads(object)

\S4method{nreads}{ChIPdata}(object)
}
\arguments{
\item{object}{a \code{ChIPdata} object}
}
\value{
the number of aligned reads in the experiment
}
\description{
Returns the number of reads in the experiment
}
