% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_SCC.R
\name{NSC}
\alias{NSC}
\title{Calculates the Normalized Strand Cross-Correlation from the SCC curve}
\usage{
NSC(scc, readLength = NULL)
}
\arguments{
\item{scc}{The tibble with the SCC curve calculate by \code{SCC}}

\item{readLength}{The readlength of the experiment. If this parameter is not present, \code{NSC} will
just calculate the ratio between the max and min of the SCC curve}
}
\value{
A numeric values
}
\description{
Calculates the NSC defined as the ratio between max. and min. of the SCC
}
