% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_pbc.R
\name{PBC}
\alias{PBC}
\title{Calculates the PCR bottleneck coefficient}
\usage{
PBC(chipdata)
}
\arguments{
\item{chipdata}{chipdata object}
}
\value{
Value between 0 and 1 with the PBC
}
\description{
Calculates the PCR bottleneck coefficient defined as:
PBC = [# of positions with exactly 1 read mapped] / [# of positions with 1 or more reads mapped ]
}
