% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{possible_chrom_sizes}
\alias{possible_chrom_sizes}
\title{Returns the chromosome sizes available in the package}
\usage{
possible_chrom_sizes()
}
\value{
A character vector
}
\description{
Returns the chromosome sizes stored in the \code{inst/extdata/chrom.size} directory of the ChIPUtils package
}
\examples{
possible_chrom_sizes()
}

