% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{possibleChrSizes}
\alias{possibleChrSizes}
\title{Returns the chromosome sizes available in the package}
\usage{
possibleChrSizes()
}
\value{
A character vector
}
\description{
Returns the chromosome sizes stored in the \code{inst/extdata/chrom.size} directory of the ChIPUtils package
}
\examples{

possibleChrSizes()
}
