% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{reads-class}
\alias{reads-class}
\title{reads class description}
\description{
Contains the reads obtained in a ChIP - seq experiment separated by strand and then by chromosome. It has one component for each strand which are object of the data.table class with a match column to identify the regions
}
\section{Slots}{

\describe{
\item{\code{readsF}}{List of data.table objects containing the reads of the ChIP - Seq experiment that have + strand.}

\item{\code{readsR}}{List of data.table object containning the reads of the ChIP - Seq experiment that have - strand}
}}
\seealso{
\code{\link{loadReads}}
}

