% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods_chipdata.R
\docType{methods}
\name{reads}
\alias{reads}
\alias{reads,ChIPdata-method}
\alias{reads}
\title{reads methods}
\usage{
reads(object)

\S4method{reads}{ChIPdata}(object)
}
\arguments{
\item{object}{A \code{ChIPdata} object}
}
\value{
The aligned reads of the experiment as \code{GAlignments} or 
\code{GAlignmentPairs} depending on if the reads are SE or PE
}
\description{
reads returns either a \code{GAlignments} or \code{GAlignmentPairs} object
depending on if the reads are SE or PE
}
\seealso{
\code{\link{ChIPdata-class}}
}
