% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods_reads.R
\docType{methods}
\name{readsF}
\alias{readsF}
\alias{readsF,reads-method}
\alias{readsF<-}
\alias{readsF<-,reads,list-method}
\title{readsF methods}
\usage{
readsF(object)

readsF(object) <- value

\S4method{readsF}{reads}(object)

\S4method{readsF}{reads,list}(object) <- value
}
\arguments{
\item{object}{A \code{reads} object}

\item{value}{A list of data.table with same format as a GenomicRanges object with an additional match column}
}
\value{
A list of data.table with same format as a GenomicRanges object with an additional match column

A reads object with modified forward reads
}
\description{
readsF returns a list of the fowward aligned reads

readsF<- assisgn a list of forward reads to a reads object
}
\examples{
file <- system.file("extdata","example",
  "encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
rr <- create_reads(file)
readsF(rr)
}
\seealso{
\code{\link{reads-class}}
}

