% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods_reads.R
\docType{methods}
\name{readsFile}
\alias{readsFile}
\alias{readsFile,reads-method}
\title{readsFile methods}
\usage{
readsFile(object)

\S4method{readsFile}{reads}(object)
}
\arguments{
\item{object}{A \code{reads} object}
}
\value{
A character with the name of the file with the aligned reads
}
\description{
readsFile returns the name of the file with the aligned reads
}
\examples{
\dontrun{
readsFile(reads)
}
}
\seealso{
\code{\link{reads-class}}
}

