% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_SCC.R
\name{RSC}
\alias{RSC}
\title{Calculates the Relative Strand Cross-Correlation from the SCC curve}
\usage{
RSC(scc, readLength)
}
\arguments{
\item{scc}{The tibble with the SCC curve calculate by \code{SCC}}

\item{readLength}{The readlength of the experiment.}
}
\value{
A numeric values
}
\description{
Calculates the RSC defined as the ratio between max. and min. of the SCC
}
