% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/base_SCC.R
\docType{methods}
\name{SCC}
\alias{SCC}
\alias{SCC,ChIPdata,numeric,logical-method}
\alias{SCC}
\title{SCC methods}
\usage{
SCC(object, shift, verbose)

\S4method{SCC}{ChIPdata,numeric,logical}(object, shift, verbose = FALSE)
}
\arguments{
\item{object}{A \code{ChIPdata} object}

\item{shift}{An non-negative integer vector}

\item{verbose}{a logical value indicating if messages should appear when calculating the SCC}
}
\value{
A \code{tibble} with two columns, shift and its respect strand cross
correlation
}
\description{
Returns a \code{tibble} with 2 columns: shift and its corresponding
strand cross correlation as defined in Kharchenko et al. 2008
}
