% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_SSD.R
\name{SSD}
\alias{SSD}
\title{Calculates the Standarized Standard Deviation}
\usage{
SSD(reads)
}
\arguments{
\item{reads}{Reads object}
}
\value{
A non-negative value
}
\description{
Calculates the SSD coefficient as defined in Planet et. al. 2011 as a uniformity measure
}
\examples{
file <- system.file("extdata","example","encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
rr <- create_reads(file)
SSD(rr)
}

