% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods_reads.R
\docType{methods}
\name{strand_cross_corr}
\alias{strand_cross_corr}
\alias{strand_cross_corr,reads,numeric,data.table-method}
\title{strand_cross_corr methods}
\usage{
strand_cross_corr(object, shift, chrom.sizes)

\S4method{strand_cross_corr}{reads,numeric,data.table}(object, shift,
  chrom.sizes)
}
\arguments{
\item{object}{A \code{reads} object}

\item{shift}{An non-negative integer vector}

\item{chrom.sizes}{An element representing the sizes of the chromosomes for the
genome of interest. It could be a \code{data.table} or a\code{character} pointing
to any of the genomes that are available in the package. The \code{data.table} must have two columns
the chromosome and its size.}
}
\value{
A \code{data.table} with two columns, shift and its respect strand cross
correlation
}
\description{
Returns a \code{data.table} with 2 columns: shift and its corresponding
strand cross correlation as defined in Kharchenko et al. 2008
}
\examples{
file <- system.file("extdata","example",
  "encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
rr <- create_reads(file)
shift <- seq(10,100,by = 10)
chrom.sizes <- data.table(read.table(system.file("inst","extdata",
  "chrom.sizes","hg19.chrom.sizes",package = "ChIPUtils")))
strand_cross_corr(rr,shift, chrom.sizes)
}

