% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods_reads.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,reads-method}
\title{summary methods}
\usage{
summary(object, ...)

\S4method{summary}{reads}(object)
}
\arguments{
\item{object}{A \code{reads} object}

\item{...}{Any other parameters of the summary method}
}
\value{
A data.table with number of reads by chromosome and by strand
}
\description{
summary returns a data.table object with the number of reads per strand and per chromosome
}
\examples{
file <- system.file("extdata","example",
  "encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
rr <- create_reads(file)
summary(rr)
}

