% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_reads.R
\name{create_reads}
\alias{create_reads}
\title{Creates a reads object}
\usage{
create_reads(readsFile, isPET = FALSE)
}
\arguments{
\item{readsFile}{Name of the bam file with the aligned reads}

\item{isPET}{Logical indicator to asses if the reads in file are PET or SET. The default value is FALSE indicating SET reads}
}
\value{
A reads object
}
\description{
Constructor of the reads class
}
\examples{
file <- system.file("extdata","example",
  "encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
create_reads(file)
}

