% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gr2dt}
\alias{gr2dt}
\title{Converts a GRanges object to data.table}
\usage{
gr2dt(gr)
}
\arguments{
\item{gr}{A GRanges object}
}
\value{
A data.table object
}
\description{
Converts a GRanges object to data.table, the columns
are gonna have the same names as the GRanges fields
}
\examples{

gr <- GRanges(seqnames = "chr1",
        ranges = IRanges(start = 3:10,width = 30),
        strand = "*")
gr2dt(gr)

}
\seealso{
\code{\link{dt2gr}}
}

