% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods_reads.R
\docType{methods}
\name{isPET}
\alias{isPET}
\alias{isPET,reads-method}
\title{isPET methods}
\usage{
isPET(object)

\S4method{isPET}{reads}(object)
}
\arguments{
\item{object}{A \code{reads} object}
}
\value{
A logical value indicating if the reads are PET (TRUE) or SET (FALSE)
}
\description{
isPET returns a logical value indicating if the object isPET
}
\examples{
file <- system.file("extdata","example",
  "encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
rr <- create_reads(file)
isPET(rr)
}
\seealso{
\code{\link{reads-class}}
}

