% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_pbc.R
\name{PBC}
\alias{PBC}
\title{Calculates the PCR bottleneck coefficient}
\usage{
PBC(reads)
}
\arguments{
\item{reads}{Reads object}
}
\value{
Value between 0 and 1 with the PBC
}
\description{
Calculates the PCR bottleneck coefficient defined as:
PBC = [# of positions with exactly 1 read mapped] / [# of positions with 1 or more reads mapped ]
}
\examples{
file <- system.file("extdata","example",
  "encode_K562_Ctcf_first3chr_Rep1.sort.bam",package = "ChIPUtils")
rr <- create_reads(file)
PBC(rr)
}

